perf: lazy load all protected route components for code splitting
Move component code from 8 protected route files into .lazy.tsx counterparts
so TanStack Router can code-split them into separate chunks, reducing the
initial bundle size.
